move log output
